+++ DSCResources/MSFT_SqlSetup/MSFT_SqlSetup.psm1
@@ -1420,8 +1420,9 @@ function Set-TargetResource
         elseif ($processExitCode -ne 0)
         {
             $setupExitMessageError = ('{0} {1}' -f $setupExitMessage, ($script:localizedData.SetupFailed))
-
             Write-Warning $setupExitMessageError
+
+            $setupEndedInError = $true
         }
         else
         {
@@ -1436,12 +1437,34 @@ function Set-TargetResource
             {
                 Write-Verbose -Message $script:localizedData.Reboot
 
+                # Rebooting, so no point in refreshing the session.
+                $forceReloadPowerShellModule = $false
+
                 $global:DSCMachineStatus = 1
             }
             else
             {
                 Write-Verbose -Message $script:localizedData.SuppressReboot
+                $forceReloadPowerShellModule = $true
             }
+        }
+        else
+        {
+            $forceReloadPowerShellModule = $true
+        }
+
+        if ((-not $setupEndedInError) -and $forceReloadPowerShellModule)
+        {
+            <#
+                Force reload of SQLPS module in case a newer version of
+                SQL Server was installed that contains a newer version
+                of the SQLPS module, although if SqlServer module exist
+                on the target node, that will be used regardless.
+                This is to make sure we use the latest SQLPS module that
+                matches the latest assemblies in GAC, mitigating for example
+                issue #1151.
+            #>
+            Import-SQLPSModule -Force
         }
 
         if (-not (Test-TargetResource @PSBoundParameters))